MCP/solutions/requirements.txt
mcp>=0.1.0
pymupdf>=1.21.1
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10
langsmith>=0.0.65
faiss-cpu>=1.10.0
python-dotenv>=1.0.0
